For the bunny prisoner locating question, I attempted to implement a totally mathematical solution to try and make the function a one liner.
I was able to figure out how to get the value of (x,1): (x**2 + x)/2 and the value of (1,y): y**2-y+2, but I was not able to figure out how to get the two to work together.

I figured out the formula of the summation being performed in my while loop where it begins shifting cells upwards was (2xy + y**2 - y)/2, but was not able to get that to work properly. So in the end I had to submit it with a while loop shifting the cells upwards.